+++ somaticseq/vcf_modifier/split_vcf.py
@@ -14,10 +14,10 @@ import uuid
 from copy import copy
 
 import somaticseq.genomic_file_parsers.genomic_file_handlers as genome
+from somaticseq.vcf_modifier.bed_util import vcfsorter
 from somaticseq.vcf_modifier.complex2indel import (
     resolve_complex_variants_into_snvs_and_indels,
 )
-from somaticseq.vcf_modifier.vcfIntersector import vcfsorter
 
 
 def run() -> argparse.Namespace:
